refactor: decouple sd-jwt verification for reusability
